refactor Page
